+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPR1AQT56TFNDV3RPPDEVPXJ"
+        "01KPR833XX4MSP2AWQ3VPG188N"
     )
     r"""The unique identifier of the data source"""