+++ packages/orq-rc/RELEASES.md
@@ -7179,3 +7179,13 @@ Based on:
 - [python v4.8.0-rc.40] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.40] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.40 - packages/orq-rc
+
+## 2026-04-21 14:48:52
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.42] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.42 - packages/orq-rc